Tests moved to appropriate namespace and converted to Dhii form
