+++ test/unit/EventManagerInterfaceTest.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace Psr\EventManager\FuncTest;
+namespace Dhii\EventManager\UnitTest;
 
-use Psr\EventManager\EventInterface as TestSubject;
+use Dhii\EventManager\EventManagerInterface as TestSubject;
 use Xpmock\TestCase;
 use Exception as RootException;
 use PHPUnit_Framework_MockObject_MockObject as MockObject;
@@ -13,14 +13,14 @@ use PHPUnit_Framework_MockObject_MockBuilder as MockBuilder;
  *
  * @since [*next-version*]
  */
-class EventInterfaceTest extends TestCase
+class EventManagerInterfaceTest extends TestCase
 {
     /**
      * The class name of the test subject.
      *
      * @since [*next-version*]
      */
-    const TEST_SUBJECT_CLASSNAME = 'Psr\EventManager\EventInterface';
+    const TEST_SUBJECT_CLASSNAME = 'Dhii\EventManager\EventManagerInterface';
 
     /**
      * Creates a new instance of the test subject.
@@ -34,16 +34,10 @@ class EventInterfaceTest extends TestCase
     public function createInstance($methods = array())
     {
         is_array($methods) && $methods = $this->mergeValues($methods, array(
-            'getName',
-            'getTarget',
-            'getParams',
-            'getParam',
-            'setName',
-            'setTarget',
-            'setParams',
-            'setParam',
-            'stopPropagation',
-            'isPropagationStopped',
+            'attach',
+            'detach',
+            'clearListeners',
+            'trigger',
         ));
 
         $mock = $this->getMockBuilder(static::TEST_SUBJECT_CLASSNAME)
